+++ gwinferno/preprocess/conversions.py
@@ -2,13 +2,7 @@
 a module that stores helpful conversion functions between parameters
 """
 
-import jax.numpy as jnp
 import numpy as np
-import xarray as xr
-from tqdm import trange
-
-from .priors import chi_effective_prior_from_isotropic_spins
-from .priors import joint_prior_from_isotropic_spins
 
 
 def chieff_from_q_component_spins(q, a1, a2, ct1, ct2):
@@ -87,117 +81,3 @@ def alpha_beta_from_mu_var(mu, var, amax=1):
     alpha = (mu**2 * (1 - mu) - mu * var) / var
     beta = (mu * (1 - mu) ** 2 - (1 - mu) * var) / var
     return alpha, beta
-
-
-def convert_component_spins_to_chieff(dat_array, param_names, injections=False):
-
-    chip = True if "chi_p" in param_names else False
-
-    q = dat_array.sel(param="mass_ratio").values
-    a_1 = dat_array.sel(param="a_1").values
-    a_2 = dat_array.sel(param="a_2").values
-    tilt_1 = dat_array.sel(param="cos_tilt_1").values
-    tilt_2 = dat_array.sel(param="cos_tilt_2").values
-    prior = dat_array.sel(param="prior").values
-
-    chi_eff = chieff_from_q_component_spins(
-        q,
-        a_1,
-        a_2,
-        tilt_1,
-        tilt_2,
-    )
-    if chip:
-        chi_p = chip_from_q_component_spins(
-            q,
-            a_1,
-            a_2,
-            tilt_1,
-            tilt_2,
-        )
-
-    new_prior = np.zeros_like(prior)
-    for ii in trange(new_prior.shape[0]):
-        for jj in range(new_prior.shape[1]):
-            if chip:
-                new_prior[ii][jj] = (
-                    prior[ii][jj]
-                    / ((2 * jnp.pi * a_1[ii][jj] ** 2) * (2 * jnp.pi * a_2[ii][jj] ** 2))
-                    * jnp.asarray(
-                        joint_prior_from_isotropic_spins(
-                            np.array(q[ii][jj]),
-                            1.0,
-                            np.array(chi_eff[ii][jj]),
-                            np.array(chi_p[ii][jj]),
-                        )
-                    )
-                )
-            else:
-                new_prior[ii][jj] = (
-                    prior[ii][jj]
-                    / ((2 * jnp.pi * a_1[ii][jj] ** 2) * (2 * jnp.pi * a_2[ii][jj] ** 2))
-                    * jnp.asarray(
-                        chi_effective_prior_from_isotropic_spins(
-                            q[ii][jj],
-                            1.0,
-                            chi_eff[ii][jj],
-                        )
-                    )[0]
-                )
-
-    new_arrays = []
-
-    if injections:
-        chi_eff_array = xr.DataArray(
-            chi_eff.reshape(1, chi_eff.shape[0]),
-            dims=["param", "injection"],
-            coords={"param": ["chi_eff"], "injection": np.arange(dat_array.injection.shape[0])},
-        )
-        prior_array = xr.DataArray(
-            new_prior.reshape(1, chi_eff.shape[0]),
-            dims=["param", "injection"],
-            coords={"param": ["prior"], "injection": np.arange(dat_array.injection.shape[0])},
-        )
-        new_arrays.append(chi_eff_array)
-        new_arrays.append(prior_array)
-
-        if chip:
-            chip_array = xr.DataArray(
-                chi_p.reshape(1, chi_p.shape[0]),
-                dims=["param", "injection"],
-                coords={"param": ["chi_p"], "injection": np.arange(dat_array.injection.shape[0])},
-            )
-            new_arrays.append(chip_array)
-
-    else:
-        chi_eff_array = xr.DataArray(
-            chi_eff.reshape(chi_eff.shape[0], 1, chi_eff.shape[1]),
-            dims=[
-                "event",
-                "param",
-                "samples",
-            ],
-            coords={"event": dat_array.event, "param": ["chi_eff"], "samples": dat_array.samples},
-        )
-        prior_array = xr.DataArray(
-            new_prior.reshape(prior.shape[0], 1, prior.shape[1]),
-            dims=["event", "param", "samples"],
-            coords={"event": dat_array.event, "param": ["prior"], "samples": dat_array.samples},
-        )
-        new_arrays.append(chi_eff_array)
-        new_arrays.append(prior_array)
-
-        if chip:
-            chip_array = xr.DataArray(
-                chi_p.reshape(1, chi_p.shape[0]),
-                dims=["event", "param", "samples"],
-                coords={"event": dat_array.event, "param": ["chi_p"], "samples": dat_array.samples},
-            )
-            new_arrays.append(chip_array)
-
-    new_dat_array = dat_array.drop_sel(param="prior")
-
-    for arr in new_arrays:
-        new_dat_array = xr.concat([new_dat_array, arr], dim="param")
-
-    return new_dat_array
